[#1087] Refactored modal
