+++ frontend/src/modal/dialogs/ConfirmDialog.less
