+++ frontend/src/modal/portal.tsx
@@ -1,8 +1,12 @@
 import * as React from 'react';
 import * as DOM from 'react-dom';
+import * as request from 'superagent';
 
-import ConfirmDialog from './ConfirmDialog.react';
+import UndoEnabledDialog from './UndoEnabledDialog';
+import UndoDisabledDialog from './UndoDisabledDialog';
+import ConfirmDialog from './dialogs/ConfirmDialog.react';
 import Modal from './Modal.react';
+import * as WpApi from '../../services/WpApi';
 
 var portalNode;
 
@@ -29,6 +33,24 @@ export function confirmDialog(title: React.ReactNode, body: React.ReactNode, okH
       <ConfirmDialog message={body} {...options} />
     </Modal>
   );
+}
+
+export function revertDialog(title: React.ReactNode, okHandler: Function) {
+  const req = WpApi
+    .get('can-revert')
+    .end((err: any, res: request.Response) => {
+      const data = res.body.data as VpApi.CanRevertResponse;
+      if (data === true) {
+        const body = <UndoEnabledDialog />;
+        portal.confirmDialog(title, body, okHandler, () => {}, {});
+      } else {
+        const body = <UndoDisabledDialog />;
+        portal.confirmDialog(title, body, () => {}, () => {}, {okButtonClasses: 'disabled'});
+      }
+    });
+
+  const cancelHandler = () => { req.abort(); };
+  portal.confirmDialog(title, '', () => {}, cancelHandler, {isLoading: true});
 }
 
 export function openPortal(children) {
